Multiple corrections to code to suppress gcc warnings - identified by Tzafrir Cohen.
git-svn-id: svn://svn.code.sf.net/p/oreka/svn/trunk@649 09dcff7a-b715-0410-9601-b79a96267cd0

+++ orkbasecxx/BatchProcessing.cpp
@@ -320,7 +320,7 @@ void BatchProcessing::ThreadHandler(void *args)
 					
 					if(CONFIG.m_transcodingSleepEveryNumFrames > 0 && CONFIG.m_transcodingSleepUs > 0)
 					{
-						if(frameSleepCounter >= CONFIG.m_transcodingSleepEveryNumFrames)
+						if(frameSleepCounter >= (unsigned int)CONFIG.m_transcodingSleepEveryNumFrames)
 						{
 							frameSleepCounter = 0;
 							struct timespec ts;
